+++ src/Component/Sort/InsertionSort/InsertionSortBars.tsx
@@ -1,18 +1,25 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { bar } from './util';
+import bar from '../../../util';
 // import { bar } from './util';
 
-const boxStyle = {
-  width: '100px',
-  height: '100px',
-  'background-color': 'white',
-  'align-text': 'center',
-  'color': 'black',
-};
 
 function InsertionSortBar({ graphBars }: any): JSX.Element {
-  const inserttionSortBars = graphBars.map((graphBar: bar) => <div style={boxStyle} key={graphBar.key.toString()}><p>{graphBar.value}</p></div>);
+  const inserttionSortBars = graphBars.map((graphBar: bar) => (
+    <div
+      style={{
+        width: '100px',
+        height: `${(graphBar.value * 50).toString()}px`,
+        backgroundColor: 'white',
+        textAlign: 'center',
+        color: 'black',
+        border: '1px solid black',
+      }}
+      key={graphBar.key.toString()}
+    >
+      <p>{graphBar.value}</p>
+    </div>
+  ));
   const flex = {
     display: 'flex',
     'flex-wrap': 'wrap',
